+++ macro/hvenh.py
@@ -1,5 +1,3 @@
-import time
-import sys
 from tkinter import *
 import pyautogui as pyauto
 import pyscreeze
@@ -16,7 +14,6 @@ frame_root = []
 btn_start = []
 val_sidestep = 0
 
-
 difficulty = "Hazardous Valley (Easy)"
 dungeonList = [
   "Hazardous Valley (Hard)",
@@ -24,7 +21,7 @@ dungeonList = [
   "Hazardous Valley (Easy)"
 ]
 
-def initialize(frame, btn, diff, runs=1):
+def initialize(frame, btn, runs, diff):
   global frame_root
   frame_root = frame
 
@@ -37,7 +34,9 @@ def initialize(frame, btn, diff, runs=1):
   shortcut.add_hotkey(util.HOTKEY_TERMINATE, util.terminate)
   btn_start.config(state=util.STATE_DISABLED)
   frame_root.update()
+
   run_dungeon(runs)
+
   btn_start.config(state=util.STATE_NORMAL)
   frame_root.update()
 
@@ -532,7 +531,7 @@ def move_to_box():
   util.move(800, 380)
   util.do_fade(0.5)
 
-def run_dungeon(runs=1):
+def run_dungeon(runs):
   run_counter = 0
   while run_counter < runs:
     util.set_reset_status(False)