+++ macro/pca.py
@@ -1,5 +1,3 @@
-import time
-import sys
 from tkinter import *
 import pyautogui as pyauto
 import pyscreeze
@@ -18,7 +16,7 @@ btn_start = []
 # UNIQUE VARIABLES
 val_sidestep = 0
 
-def initialize(frame, btn, runs=1):
+def initialize(frame, btn, runs):
   global frame_root
   frame_root = frame
 
@@ -207,7 +205,7 @@ def position_third_shadow():
   util.move(690, 150)
   util.do_dash(1)
 
-def run_dungeon(runs=1):
+def run_dungeon(runs):
   run_counter = 0
   while run_counter < runs:
     util.set_reset_status(False)