+++ macro/sca.py
@@ -1,5 +1,3 @@
-import time
-import sys
 from tkinter import *
 import pyautogui as pyauto
 import pyscreeze
@@ -16,7 +14,7 @@ frame_root = []
 btn_start = []
 portal_counter = 0
 
-def initialize(frame, btn, runs=1):
+def initialize(frame, btn, runs):
   global frame_root
   frame_root = frame
 
@@ -26,7 +24,9 @@ def initialize(frame, btn, runs=1):
   shortcut.add_hotkey(util.HOTKEY_TERMINATE, util.terminate)
   btn_start.config(state=util.STATE_DISABLED)
   frame_root.update()
+
   run_dungeon(runs)
+
   btn_start.config(state=util.STATE_NORMAL)
   frame_root.update()
 
@@ -266,7 +266,7 @@ def path_find(unit):
     util.do_dash(1)
     util.do_fade(0.1)
 
-def run_dungeon(runs=1):
+def run_dungeon(runs):
   global portal_counter
   run_counter = 0
   while run_counter < runs: